// src/app/interfaces/movieImages.ts
// Generated by https://quicktype.io

export interface MovieImages {
    backdrops: Backdrop[];
}

export interface Backdrop {
    aspect_ratio: number;
    height:       number;
    iso_639_1:    null;
    file_path:    string;
    vote_average: number;
    vote_count:   number;
    width:        number;
}

export interface BackdropFinalUrl {
    backdropUrl: string
}